third_party/blink/web_tests/svg/hittest/update-rect-expected.txt

Regression test for http://crbug.com/436214. This tests updating SVG <rect> elements having various stroke styles. If this test passes, you will see "PASS" below.

PASS

(400, 118)
(460, 180)
(263, 83)
(200, 62)
(235, 112)
(160, 350)
(233, 413)
(160, 300)
(400, 350)
(350, 370)
(200, 230)
(267, 236)
(304, 218)
(307, 266)
(707, 65)
(710, 130)
(712, 164)
(712, 216)
(712, 264)
(712, 316)
(610, 425)
(739, 394)
